requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask_JWT_Extended==4.0.0
itsdangerous==1.1.0
MarkupSafe==1.1.1
pymongo==3.10.1
Werkzeug==1.0.1
Flask-Bcrypt==0.7.1
PyJWT==2.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
pika==1.3.2